scripts/requirements.txt
pillow
pywavefront
